zeroPadding works!!! ... i think
I shall now change names for it to make sense but it doesn't give any errors so should be all good.

Explaination: So the first part is a for loop so that the vectors can be changed into a row vector in case its a column vector. Then it passes into the if statement which runs through which vector is bigger and then adding zeros to the shorter vectors.